Refactor org formatter code for formatting logs.
